Add files via upload
What is contained in this folder is a summary of the work and study I have put into a group project that involved creating a model capable of:

Given an input image of a face as a query, the model had to extract 10 images from a given gallery that featured the same subject as the input image, serving as the search key.

The uploaded code (due to space constraints) is partial in the sense that it lacks the corresponding folders where various datasets were stored, which we used to train the models. The files containing the models themselves are also absent because they are quite large.

The scripts are organized in a "notebook" style, which simplified debugging and sharing of materials among the members of the workgroup.

The four files present above are respectively the three scripts used to train three of the most representative models we trained, along with their functions for constructing custom dataloaders. The fourth file, "utils," contains some of the functions we used to customize, clean, and manage the various datasets we used.